Update download link to latest (2.1.669)
I'm not sure if this is desired but throwing up a PR to reduce maintenance if wanted. You can link to the latest release with a slightly different format: `https://github.com/KirillOsenkov/MSBuildStructuredLog/releases/latest/download/MSBuildStructuredLogSetup.exe` - this will always grab the latest release for that repo. No worries if unwanted and a specific version is desired here - just close away :)